scripts/requirements.txt
kafka-python
pandas
xgboost
torch
joblib
scikit-learn
numpy
fastapi
uvicorn
pyarrow
pyyaml